PHP files
This is still a working project. I have uploaded all files except for the MySQL connect page for obvious reasons. I decided to play around with a new method for organizing and displaying the mark up, which I find keeps the work a little easier to make changes to the front end and makes it easier to print large blocks of html. You can see how I have set everything up in the master.php file.